requirements.txt
flask
functions-framework==3.*
google-genai
